Use cloud basePath for API clients from configuration passed in
The configuration for API clients isn't using the cloud configuration value to derive basePath. However, the auth code uses the cloud value to derive base path.

This causes problems when the cloud configuration value isn't the same as the default used in the API configuration. This causes API failures since the access token is provisioned from a different endpoint.

This derives a basePath for API client configuration using the cloud configuration value passed in so they are in sync.